Set of change, update of version new handler classes for queries spliterd types
