+++ src/TheLizzards.Data.Azure/TheLizzards.Data.Azure.csproj
@@ -3,8 +3,8 @@
 		<Description>Default implementatin of differenet services in Azure.</Description>
 		<Copyright>TheLizzards 2016 - 2017</Copyright>
 		<AssemblyTitle>TheLizzards.Data.Azure</AssemblyTitle>
-		<VersionPrefix>2.0.0</VersionPrefix>
-		<Version>2.0.0</Version>
+		<VersionPrefix>2.1.0-beta4</VersionPrefix>
+		<Version>2.1.0-beta4</Version>
 		<AssemblyName>TheLizzards.Data.Azure</AssemblyName>
 		<PackageTags>azure;thelizzards</PackageTags>
 		<RepositoryType>git</RepositoryType>
