+++ src/TheLizzards.Data/CQRS/Queries/IWithEmail.cs
@@ -1,0 +1,9 @@
+﻿using TheLizzards.Data.Types.Entites;
+
+namespace TheLizzards.Data.CQRS.Queries
+{
+	public interface IWithEmail<TResult>
+	{
+		TResult WithEmail(Email email);
+	}
+}
