+++ src/TheLizzards.Data/CQRS/Queries/IWithName.cs
@@ -1,0 +1,7 @@
+﻿namespace TheLizzards.Data.CQRS.Queries
+{
+	public interface IWithName<TResult>
+	{
+		TResult WithName(string name);
+	}
+}
