+++ src/TheLizzards.Data/Claims/IClaimsProvider.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Security.Claims;
 
-namespace TheLizzards.Data.Types.Contracts
+namespace TheLizzards.Data.Claims
 {
 	public interface IClaimsProvider
 	{
