+++ src/TheLizzards.Data/DDD/IWithEmailIndex.cs
@@ -1,0 +1,9 @@
+﻿using TheLizzards.Data.Types.Entites;
+
+namespace TheLizzards.Data.DDD
+{
+	public interface IWithEmailIndex
+	{
+		Email Email { get; }
+	}
+}
