+++ src/TheLizzards.Data/DDD/IWithNameIndex.cs
@@ -1,0 +1,7 @@
+﻿namespace TheLizzards.Data.DDD
+{
+	public interface IWithNameIndex
+	{
+		string Name { get; }
+	}
+}
