+++ src/TheLizzards.Data/TheLizzards.Data.csproj
@@ -3,8 +3,8 @@
 		<Description>Implementation of common data access patterns.</Description>
 		<Copyright>TheLizzards 2016 - 2017</Copyright>
 		<AssemblyTitle>Lizzards Data</AssemblyTitle>
-		<VersionPrefix>2.0.1-beta3</VersionPrefix>
-		<Version>2.0.1-beta3</Version>
+		<VersionPrefix>2.1.0-beta4</VersionPrefix>
+		<Version>2.1.0-beta4</Version>
 		<AssemblyName>TheLizzards.Data</AssemblyName>
 		<PackageTags>data;thelizzards</PackageTags>
 		<RepositoryType>git</RepositoryType>
