+++ src/TheLizzards.Data/Types/Address.cs
@@ -1,4 +1,4 @@
-﻿namespace TheLizzards.Data.Types.Entites
+﻿namespace TheLizzards.Data.Types
 {
 	public sealed class Address
 	{
