+++ src/TheLizzards.Data/Types/ContatDetailsTypes.cs
@@ -1,4 +1,4 @@
-﻿namespace TheLizzards.Data.Types.Entites
+﻿namespace TheLizzards.Data.Types
 {
 	public sealed class ContatDetailsTypes
 	{
