--- src/TheLizzards.Data/Types/Contracts/IIdProvider.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace TheLizzards.Data.Types.Contracts
-{
-	public interface IIdProvider
-	{
-		Guid Id { get; }
-	}
-}
