--- src/TheLizzards.Data/Types/Contracts/INameProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace TheLizzards.Data.Types.Contracts
-{
-	public interface INameProvider
-	{
-		string Name { get; }
-	}
-}
