+++ src/TheLizzards.Mvc.Data/TheLizzards.Mvc.Data.csproj
@@ -3,8 +3,8 @@
 		<Description>Implementation of common data access patterns.</Description>
 		<Copyright>TheLizzards 2016 - 2017</Copyright>
 		<AssemblyTitle>Lizzards Data for Mvc</AssemblyTitle>
-		<VersionPrefix>1.1.3-beta3</VersionPrefix>
-		<Version>1.1.3-beta3</Version>
+		<VersionPrefix>1.1.3-beta4</VersionPrefix>
+		<Version>1.1.3-beta4</Version>
 		<TargetFramework>netstandard1.6</TargetFramework>
 		<AssemblyName>TheLizzards.Mvc.Data</AssemblyName>
 		<PackageTags>data;thelizzards</PackageTags>
